Removed Tester views that will no longer be used Abstracted out Optimization Strategies to be an inheritance structure. Changed logic so that if you generate schedule with no classes it now removes them. Fixed issue where new accounts wouldn't be able to see the index Changed CustomEventRequest object to take in DateTime instead of strings. Fixed bug where classes couldnt be removed if the user hadn't refreshed but had generated a schedule with a new course.
