+++ ucsdscheduleme/Controllers/AccountController.cs
@@ -1,12 +1,10 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using ucsdscheduleme.Models;
-using System.Security.Claims;
 
 namespace ucsdscheduleme.Controllers
 {
