+++ ucsdscheduleme/Models/CustomEvent.cs
@@ -9,8 +9,8 @@ namespace ucsdscheduleme.Models
     {
         public string Name { get; set; }
         public Days Days { get; set; }
-        public string StartTime { get; set; }
-        public string EndTime { get; set; }
+        public DateTime StartTime { get; set; }
+        public DateTime EndTime { get; set; }
     }
 
     public class CustomEventResponse
