+++ ucsdscheduleme/Repo/ScheduleRepo.cs
@@ -3,7 +3,6 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using ucsdscheduleme.Models;
-using PossibleSchedules = System.Collections.Generic.List<System.Collections.Generic.List<ucsdscheduleme.Models.Section>>;
 
 namespace ucsdscheduleme.Repo
 {
@@ -239,413 +238,6 @@ namespace ucsdscheduleme.Repo
             return true;
         }
 
-        /// <summary>
-        /// Returns the schedule with the highest received GPAs from CAPEs.
-        /// </summary>
-        /// <returns>The schedule with highest CAPEs GPA</returns>
-        /// <param name="possibleSchedules">Possible schedules with no time conflicts.</param>
-        public List<Section> HighestGPA(List<List<Section>> possibleSchedules)
-        {
-            List<Section> result = possibleSchedules[0];
-            decimal highestGPA = 0;
-
-            foreach (List<Section> schedule in possibleSchedules)
-            {
-                decimal currentGPA = 0;
-
-                foreach (Section section in schedule)
-                {
-                    // Find a cape that corresponds to the course and professor
-                    var cape = section.Course.Cape.First(c => c.ProfessorId == section.Professor.Id);
-
-                    currentGPA += cape?.AverageGradeReceived ?? 0M;
-                }
-
-                // Compare GPA of this schedule to the highest schedule GPA
-                if (currentGPA > highestGPA)
-                {
-                    highestGPA = currentGPA;
-                    result = schedule;
-                }
-            }
-
-            return result;
-        }
-
-        /// <summary>
-        /// Returns the schedule with the highest overall quality from RateMyProfessor.
-        /// </summary>
-        /// <returns>The schedule with the hightest overall quality from RateMyProfessor.</returns>
-        /// <param name="possibleSchedules">Possible schedules with no time conflicts.</param>
-        public List<Section> RMPRating(List<List<Section>> possibleSchedules)
-        {
-            List<Section> result = possibleSchedules[0];
-            decimal highestRating = 0;
-
-            foreach (List<Section> schedule in possibleSchedules)
-            {
-                decimal currentRating = 0;
-
-                foreach (Section section in schedule)
-                {
-                    currentRating += section.Professor.RateMyProfessor.OverallQuality;
-                }
-
-                // Compare GPA of this schedule to the highest schedule GPA
-                if (currentRating > highestRating)
-                {
-                    highestRating = currentRating;
-                    result = schedule;
-                }
-            }
-
-            return result;
-        }
-
-        /// <summary>
-        /// Returns the schedule with the earliest ending time.
-        /// </summary>
-        /// <returns>The schedule with earliest ending time.</returns>
-        /// <param name="possibleSchedules">Possible schedules without time conflicts.</param>
-        public List<Section> EarliestEnd(List<List<Section>> possibleSchedules)
-        {
-            List<Section> result = possibleSchedules[0];
-            DateTime? earliestEnd = new DateTime(1, 1, 1, 23, 59, 0);
-
-            // Get the latest time for a schedule
-            foreach (List<Section> schedule in possibleSchedules)
-            {
-                DateTime? lastestScheduleTime = new DateTime(1, 1, 1, 0, 0, 0);
-
-                foreach (Section section in schedule)
-                {
-                    DateTime? latestCourseTime = new DateTime(1, 1, 1, 0, 0, 0);
-                    foreach (Meeting meeting in section.Meetings)
-                    {
-                        if (meeting.MeetingType != MeetingType.Review &&
-                            meeting.MeetingType != MeetingType.Final &&
-                            meeting.MeetingType != MeetingType.Midterm)
-                        {
-                            // Get the latest end time for a course
-                            if (meeting.EndTime > latestCourseTime)
-                            {
-                                latestCourseTime = meeting.EndTime;
-                            }
-                        }
-                    }
-
-                    // Get the latest end time for the schedule
-                    if (latestCourseTime > lastestScheduleTime)
-                    {
-                        lastestScheduleTime = latestCourseTime;
-                    }
-                }
-
-                // Get scheduele with the earliest schedule
-                if (earliestEnd > lastestScheduleTime)
-                {
-                    earliestEnd = lastestScheduleTime;
-                    result = schedule;
-                }
-            }
-            return result;
-        }
-
-        /// <summary>
-        /// Returns the schedule with the latest start time.
-        /// </summary>
-        /// <returns>The schedule with the latest start time.</returns>
-        /// <param name="possibleSchedules">Possible schedules.</param>
-        public List<Section> LatestStart(List<List<Section>> possibleSchedules)
-        {
-            List<Section> result = possibleSchedules[0];
-            DateTime? latestStart = new DateTime(1, 1, 1, 0, 0, 0);
-
-            // Get the lastest time for a schedule
-            foreach (List<Section> schedule in possibleSchedules)
-            {
-                DateTime? earliestScheduleTime = new DateTime(1, 1, 1, 23, 59, 0);
-
-                foreach (Section section in schedule)
-                {
-                    DateTime? earliestCourseTime = new DateTime(1, 1, 1, 23, 59, 0);
-
-                    foreach (Meeting meeting in section.Meetings)
-                    {
-                        if (meeting.MeetingType != MeetingType.Review &&
-                            meeting.MeetingType != MeetingType.Final &&
-                            meeting.MeetingType != MeetingType.Midterm)
-                        {
-                            // Get earliest start time for course
-                            if (meeting.StartTime < earliestCourseTime)
-                            {
-                                earliestCourseTime = meeting.StartTime;
-                            }
-                        }
-                    }
-
-                    // Get the earliest start time for schedule
-                    if (earliestCourseTime < earliestScheduleTime)
-                    {
-                        earliestScheduleTime = earliestCourseTime;
-                    }
-                }
-                
-                // Compare
-                if (latestStart < earliestScheduleTime)
-                {
-                    latestStart = earliestScheduleTime;
-                    result = schedule;
-                }
-            }
-            return result;
-        }
-
-        /// <summary>
-        /// Returns the schedule with the least days scheduled.
-        /// </summary>
-        /// <returns>The schedule with the least days.</returns>
-        /// <param name="possibleSchedules">Possible schedules with no time conflicts.</param>
-        public List<Section> LeastDays(List<List<Section>> possibleSchedules)
-        {
-            int leastDay = 10;
-            List<Section> leastDaySchedule = possibleSchedules[0];
-
-            // Iterate through the possible schedules and counting the number of days
-            // per schedule
-            foreach (List<Section> schedule in possibleSchedules)
-            {
-                int numDays = NumDays(schedule);
-
-                // Checks to see if the current schedule has less days than the current least.
-                if (numDays < leastDay)
-                {
-                    leastDay = numDays;
-                    leastDaySchedule = schedule;
-                }
-            }
-            return leastDaySchedule;
-        }
-
-        /// <summary>
-        /// Returns the schedule with the most days scheduled.
-        /// </summary>
-        /// <returns>The schedule with the most days.</returns>
-        /// <param name="possibleSchedules">Possible schedules with no time conflicts.</param>
-        public List<Section> MostDays(List<List<Section>> possibleSchedules)
-        {
-            int mostDay = 0;
-            List<Section> mostDaySchedule = possibleSchedules[0];
-
-            // Iterate through the possible schedules and counts how many days are scheduled.
-            foreach (List<Section> schedule in possibleSchedules)
-            {
-                int numDays = NumDays(schedule);
-
-                // Compares the current schedule days to the current max days.
-                if (numDays > mostDay)
-                {
-                    mostDay = numDays;
-                    mostDaySchedule = schedule;
-                }
-            }
-            return mostDaySchedule;
-        }
-
-        /// <summary>
-        /// Calculates the number of days the schedule has classes on
-        /// </summary>
-        /// <param name="schedule">A schedule of classes</param>
-        /// <returns> the number of days in the schedule</returns>
-        private int NumDays(List<Section> schedule)
-        {
-            int numDays = 0;
-
-            var allDays = schedule.SelectMany(s => s.Meetings)
-                                  .Where(m => !FormatRepo.IsOneTimeEvent(m.MeetingType))
-                                  .Select(m => m.Days);
-
-            List<Days> days = Enum.GetValues(typeof(Days)).Cast<Days>().ToList();
-
-            foreach(var day in days)
-            {
-                if(allDays.Any(d => d.HasFlag(day)))
-                {
-                    numDays++;
-                }
-            }
-
-            return numDays; 
-        }
-
-        /// <summary>
-        /// Returns the schedule with the least gaps in between classes.
-        /// </summary>
-        /// <returns>The schedule with the least gaps.</returns>
-        /// <param name="possibleSchedules">Possible schedules with no time conflicts.</param>
-        public List<Section> LeastGaps(List<List<Section>> possibleSchedules)
-        {
-            List<Section> leastGapsSchedule = possibleSchedules[0];
-            TimeSpan leastGap = new TimeSpan(99,59,59);
-
-            // Iterate through the possible schedules and separate the times into respective days.
-            foreach (List<Section> schedule in possibleSchedules)
-            {
-                TimeSpan currGap = TotalGap(schedule);
-
-                // Checks to see if current schedule has less gaps than current least.
-                if (currGap < leastGap)
-                {
-                    leastGapsSchedule = schedule;
-                    leastGap = currGap;
-                }
-            }
-            return leastGapsSchedule;
-        }
-
-        /// <summary>
-        /// Returns the schedule with the most gaps in between classes.
-        /// </summary>
-        /// <returns>The schedule with the most gaps.</returns>
-        /// <param name="possibleSchedules">Possible schedules with no time conflicts.</param>
-        public List<Section> MostGaps(List<List<Section>> possibleSchedules)
-        {
-            List<Section> mostGapsSchedule = possibleSchedules[0];
-            TimeSpan mostGap = new TimeSpan(0,0,0);
-
-            // Iterate through the possible schedules and separate the times into respective days.
-            foreach (List<Section> schedule in possibleSchedules)
-            {
-                TimeSpan currGap = TotalGap(schedule);
-
-                // Checks to see if current schedule has less gaps than current least.
-                if (currGap > mostGap) 
-                {
-                    mostGapsSchedule = schedule;
-                    mostGap = currGap;
-                }
-            }
-            return mostGapsSchedule;
-        }
-
-        public List<Section> Optimize(Optimization optimization, PossibleSchedules schedules)
-        {
-            List<Section> schedule;
-            switch (optimization)
-            {
-                case Optimization.HighestGPA:
-                    schedule = HighestGPA(schedules);
-                    break;
-                case Optimization.HighestRMP:
-                    schedule = RMPRating(schedules);
-                    break;
-                case Optimization.EarlyEnd:
-                    schedule = EarliestEnd(schedules);
-                    break;
-                case Optimization.LateStart:
-                    schedule = LatestStart(schedules);
-                    break;
-                case Optimization.MostDays:
-                    schedule = MostDays(schedules);
-                    break;
-                case Optimization.LeastDays:
-                    schedule = LeastDays(schedules);
-                    break;
-                case Optimization.MostGaps:
-                    schedule = MostGaps(schedules);
-                    break;
-                case Optimization.LeastGaps:
-                    schedule = LeastGaps(schedules);
-                    break;
-                default:
-                    throw new FormatException();
-            }
-            return schedule;
-        }
-
-        /// <summary>
-        /// Calculates the total gap size for a schedule.
-        /// </summary>
-        /// <param name="schedule">Schedule of classes</param>
-        /// <returns>The total time gaps for a particular schedule</returns>
-        private TimeSpan TotalGap(List<Section> schedule)
-        {
-            TimeSpan gap = new TimeSpan(0);
-            List<DateTime> Monday = new List<DateTime>();
-            List<DateTime> Tuesday = new List<DateTime>();
-            List<DateTime> Wednesday = new List<DateTime>();
-            List<DateTime> Thursday = new List<DateTime>();
-            List<DateTime> Friday = new List<DateTime>();
-
-            var allMeetings = schedule.SelectMany(s => s.Meetings);
-            foreach(Meeting meeting in allMeetings)
-            {
-                if (!FormatRepo.IsOneTimeEvent(meeting.MeetingType) 
-                    && meeting.StartTime.Hour != 0 
-                    && meeting.EndTime.Hour != 0)
-                {
-                    if (meeting.Days.HasFlag(Days.Monday))
-                    {
-                        Monday.Add(meeting.StartTime);
-                        Monday.Add(meeting.EndTime);
-                    }
-                    if (meeting.Days.HasFlag(Days.Tuesday))
-                    {
-                        Tuesday.Add(meeting.StartTime);
-                        Tuesday.Add(meeting.EndTime);
-                    }
-                    if (meeting.Days.HasFlag(Days.Wednesday))
-                    {
-                        Wednesday.Add(meeting.StartTime);
-                        Wednesday.Add(meeting.EndTime);
-                    }
-                    if (meeting.Days.HasFlag(Days.Thursday))
-                    {
-                        Thursday.Add(meeting.StartTime);
-                        Thursday.Add(meeting.EndTime);
-                    }
-                    if (meeting.Days.HasFlag(Days.Friday))
-                    {
-                        Friday.Add(meeting.StartTime);
-                        Friday.Add(meeting.EndTime);
-                    }
-                }
-            }
-
-            // Sort the class times in each schedule by start time.
-            Monday.Sort();
-            Tuesday.Sort();
-            Wednesday.Sort();
-            Thursday.Sort();
-            Friday.Sort();
-            
-            // Calculate the gaps in the schedule.
-            for (int i = 1; i < Monday.Count - 1; i += 2)
-            {
-                gap += (Monday[i + 1].TimeOfDay - Monday[i].TimeOfDay);
-            }
-
-            for (int i = 1; i < Tuesday.Count - 1; i += 2)
-            {
-                gap += (Tuesday[i + 1].TimeOfDay - Tuesday[i].TimeOfDay);
-            }
-
-            for (int i = 1; i < Wednesday.Count - 1; i += 2)
-            {
-                gap += (Wednesday[i + 1].TimeOfDay - Wednesday[i].TimeOfDay);
-            }
-
-            for (int i = 1; i < Thursday.Count - 1; i += 2)
-            {
-                gap += (Thursday[i + 1].TimeOfDay - Thursday[i].TimeOfDay);
-            }
-
-            for (int i = 1; i < Friday.Count - 1; i += 2)
-            {
-                gap += (Friday[i + 1].TimeOfDay - Friday[i].TimeOfDay);
-            }
-
-            return gap;
-        }
+        
     }
 }
