--- ucsdscheduleme/Views/Schedule/Index.cshtml
@@ -1,68 +1,0 @@
-﻿@model ucsdscheduleme.Models.ScheduleViewModel
-@{
-    ViewData["Title"] = "Schedule";
-}
-
-<h1>Schedule Generator</h1>
-<div class="container">
-    <div class="col-md-4">
-    	<h3>Choose Your classes</h3>   
-        <form asp-action="GenerateSchedule">
-        	<label for="course1name" class="control-label">Course 1</label>
-            <input type="text" name="course1name" class="form-control"/>
-            <br />
-        	<label for="course2name" class="control-label">Course 2</label>
-        	<input type="text" name="course2name" class="form-control"/>
-            <br />
-        	<label for="course3name" class="control-label">Course 3</label>
-        	<input type="text" name="course3name" class="form-control"/>
-        	<br />
-            <label for="course4name" class="control-label">Course 4</label>
-            <input type="text" name="course4name" class="form-control"/>
-        	<br />
-        	<p><input type="submit" value="Generate" class="btn btn-default"/></p>
-        	<br />
-        </form>
-    </div>
-
-    <div class="col-md-4">
-        <h3>Schedule</h3>
-        <table class="table table-bordered table-hover">
-            <thead>
-    			<tr>
-    				<th scope="col">ID</th>
-    				<th scope="col">Name</th>
-    				<th scope="col">Units</th>
-					<th scope="col">Meeting Type</th>
-					<th scope="col">Start Time</th>
-					<th scope="col">End Time</th>
-    			</tr>
-			</thead>
-			@if (Model != null)
-			{
-			    foreach (var course in @Model.CourseList)
-			    {
-			        <tr>
-				        <td scope="row">@course.Id</td>
-                        <td>@course.CourseName</td>
-				        <td>@course.Units</td>
-				        @if (course.Sections != null) {
-                            foreach(var section in @course.Sections)
-    				        {
-    				            if (section.Meetings != null)
-                                {
-				                    foreach(var meeting in @section.Meetings)
-                                    {
-                                        <td>@meeting.MeetingType</td>
-                                        <td>@meeting.StartTime</td>
-                                        <td>@meeting.EndTime</td>
-                                    }
-				                }
-    				        }
-				        }
-			        </tr>
-			    }
-			}
-		</table>
-    </div>
-</div>
