--- ucsdscheduleme/Views/Scrape/Index.cshtml
@@ -1,57 +1,0 @@
-﻿@model ucsdscheduleme.Models.ScrapeViewModel
-@{
-    ViewData["Title"] = "Scrape Portal";
-}
-
-<div class="container">
-    <h4>Scrape Portal</h4>
-    <hr />
-    <div class="row">
-        <div class="col-md-4">
-            <form asp-action="ScrapeSingleCape">
-                <div asp-validation-summary="ModelOnly" class="text-danger"></div>
-                <div class="form-group">
-                    <label asp-for="Url" class="control-label"></label>
-                    <input asp-for="Url" class="form-control" />
-                    <span asp-validation-for="Url" class="text-danger"></span>
-                </div>
-                <div class="form-group">
-                    <label asp-for="ScrapeSite" class="control-label"></label>
-                    <select asp-for="ScrapeSite" class="form-control" asp-items="Html.GetEnumSelectList<ScrapeSite>()"></select>
-                    <br />    
-                    <input type="submit" value="Scrape Single Url" class="btn btn-default" />
-                </div>
-            </form>
-            <hr />
-            <div class="row">
-                @* This is currently disabled because this action isn't going to do anything for awhile.*@
-                <button asp-action="ScrapeAllCapes" class="btn btn-danger disabled">Scrape All Capes</button>
-            </div>
-        </div>
-        <div class="col-md-8">
-            @if (Model != null && Model.ScrapeResults.Count == 1)
-            {
-                var scrapeResult = Model.ScrapeResults.First();
-                <ul class="list-group">
-                    <li class="list-group-item">Instructor: @scrapeResult.InstructorName</li>
-                    <li class="list-group-item">Qtr/Yr: @scrapeResult.Term</li>
-                    <li class="list-group-item">Enrollment: @scrapeResult.StudentsEnrolled</li>
-                    <li class="list-group-item">CAPEs Returned: @scrapeResult.NumberOfEvaluation</li>
-                    <li class="list-group-item">Recommend the course: @scrapeResult.RecommendedClass</li>
-                    <li class="list-group-item">Recommend the instructor: @scrapeResult.RecommendedProfessor</li>
-                    <li class="list-group-item">Average study hours per week: @scrapeResult.StudyHoursPerWeek</li>
-                    <li class="list-group-item">Average grades students expected: @scrapeResult.AverageGradeExpected</li>
-                    <li class="list-group-item">Average grades students received: @scrapeResult.AverageGradeReceived</li>
-				    <li class="list-group-item">Overall quality: @scrapeResult.OverallQuality</li>
-				    <li class="list-group-item">Percentage would take again: @scrapeResult.WouldTakeAgain</li>
-				    <li class="list-group-item">Level of difficulty: @scrapeResult.LevelOfDifficulty</li>
-                </ul>
-            }
-            @* Else if will eventually go here, showing metadata from list of returned scrapes*@
-        </div>
-    </div>
-</div>
-
-@section Scripts {
-    @{await Html.RenderPartialAsync("_ValidationScriptsPartial");}
-}
