--- ucsdscheduleme/Views/TestSchedule/GenerateSchedule.cshtml
@@ -1,1 +1,0 @@
-﻿
