+++ ucsdscheduleme/ucsdscheduleme.csproj
@@ -16,10 +16,6 @@
 
   <ItemGroup>
     <Compile Remove="Models\ScrapeResultRateMyProf.cs" />
-    <Folder Include="Views\Schedule\" />
-  </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Views\TestSchedule\" />
   </ItemGroup>
   <ItemGroup>
     <None Remove="Views\Home\index.html" />
