+++ ucsdscheduleme/wwwroot/js/site.js
@@ -1,6 +1,5 @@
 ﻿function setup() {
     updateSchedule(myApp.courses);
-    console.log("setup()");
 }
 
 /* called when DOM is ready */
@@ -19,7 +18,6 @@ function typeAheadCallout(input) {
     xhr.open("POST", url, true);
     xhr.setRequestHeader('Content-Type', 'application/json');
     var send = { "input": input, "alreadyAddedCourses": myApp.coursesToSchedule };
-    console.log("Payload: " + JSON.stringify(send));
     xhr.send(JSON.stringify(send));
 
     // When the text is edited, it clears the search and populates it
@@ -30,7 +28,6 @@ function typeAheadCallout(input) {
 
             for (i = 0; i < text.length; i++) {
                 populateSearch(text[i]);
-                console.log(text[i]);
             }
         }
     }
@@ -70,9 +67,6 @@ function populateSearch(data) {
 
     course.id = data.id;
 
-    console.log("populateSearch");
-    console.log(course);
-
     // Add it to the drop down
     courses.append(course);
 }
@@ -107,7 +101,6 @@ function addToScheduleList(event, isCustom = false) {
     list.append(course);
 
     // Add course to current schedule
-    console.log("data.id: " + event.id);
     myApp.coursesToSchedule.push(event.id);
 }
 
@@ -116,7 +109,6 @@ function addToScheduleList(event, isCustom = false) {
  * @param {HTMLElement} data class div selected to add
  */
 function addCourse(data) {
-    console.log("Data: " + JSON.stringify(data));
 
     // Hide dropdown menu
     var dropdown = document.getElementById("courseItems");
@@ -135,7 +127,6 @@ function removeCustomEventCallout(courseId) {
     var url = myApp.urls.removeCustomEvent + "?courseId=" + courseId;
     xhr.open("DELETE", url, true);
     xhr.setRequestHeader('Content-Type', 'application/json');
-    console.log("Payload: " + JSON.stringify(courseId));
     xhr.send();
 }
 
@@ -147,7 +138,13 @@ function removeCourse(e) {
     var course = e.target.parentNode.parentNode;
     var id = course.id;
 
-    var index = myApp.coursesToSchedule.indexOf(Number(id));
+    var index = -1;
+    for (var i = 0; i < myApp.coursesToSchedule.length; i++) {
+        if (myApp.coursesToSchedule[i] == id) {
+            index = i;
+            break;
+        }
+    }
 
     if (index >= 0) {
         myApp.coursesToSchedule.splice(index, 1);
@@ -438,12 +435,9 @@ function showBaseAndAllSections(ids) {
     }
 
     var sectionsKeys = Object.keys(sectionEvents);
-    console.log("showBaseAndAllSections key: " + JSON.stringify(sectionsKeys));
     sectionsKeys.forEach(function (key) {
         var section = sectionEvents[key];
-        console.log("showBaseAndAllSections section: " + JSON.stringify(section));
         for(var sectionIndex in section) {
-            console.log("showBaseAndAllSections section[sectionIndex]: " + JSON.stringify(section[sectionIndex]));
             var eventElement = insertMeeting(section[sectionIndex], ids.courseId, ids.baseId, key);
             eventElement.classList.add("event-activated");
         }
@@ -455,27 +449,19 @@ function showBaseAndAllSections(ids) {
 
 
 function changeSchedule(event) {
-    console.log("changeSchedule event")
-    console.log(event)
-
     var info = extractEventInfo(event);
-    console.log("changeSchedule info")
-    console.log(info)
 
     // base selected
     if (info.isBaseEvent) {
-        console.log("changeSchedule calls showAllBasesAndAllSections")
         showAllBasesAndAllSections(info);
     }
     // section selected
     else {
-        console.log("changeSchedule calls showBaseAndAllSections")
         showBaseAndAllSections(info);
     }
 }
 
 function changeScheduleSectionCallout(oldSectionId, newSectionId) {
-    console.log("changeScheduleSectionCallout");
     var xhr = new XMLHttpRequest();
     var url = myApp.urls.changeScheduleSection;
     xhr.open("POST", url, true);
@@ -502,7 +488,6 @@ function updateSelectedSection(event) {
     updateSchedule(myApp.courses);
 
     isEditing = false;
-    console.log("before changeScheduleSectionCallout");
     changeScheduleSectionCallout(currentSectionId, ids.sectionId);
     showEditButtons();
 }
@@ -548,7 +533,6 @@ function updatedSelectedSectionAsBase(event) {
     updateMeetings(myApp.courses);
 
     isEditing = false;
-    console.log("before changeScheduleSectionCallout");
     changeScheduleSectionCallout(currentSectionId, ids.sectionId);
     showEditButtons();
 }
@@ -568,18 +552,10 @@ function updateEvent(event) {
 
 function activateSelectedBasesAndSections(event) {
     var infoForSelected = extractEventInfo(event, false);
-    
-    console.log("activateSelectedBasesAndSections")
-    console.log(infoForSelected)
 
     var allActivatedEvents = document.getElementsByClassName('event-activated');
-    console.log("allActivatedEvents")
-    console.log(allActivatedEvents)
-
 
     if (!infoForSelected.isBaseEvent) {
-
-        console.log("activateSelectedBasesAndSections !infoForSelected.isBaseEvent")
         var toDeactivate = [];
         for (var j = 0; j < allActivatedEvents.length; j++) {
             var classList = allActivatedEvents[j].classList;
@@ -588,11 +564,9 @@ function activateSelectedBasesAndSections(event) {
             if (classList.contains(infoForSelected.courseId) && classList.contains(infoForSelected.baseId)) {
                 if (classList.contains(infoForSelected.sectionId) || infoForCurrent.isBaseEvent) {
                 } else {
-                    console.log("Remove 1");
                     toDeactivate.push(allActivatedEvents[j]);
                 }
             } else {
-                console.log("Remove 2");
                 toDeactivate.push(allActivatedEvents[j]);
             }
         }
@@ -966,9 +940,6 @@ function updateOneTimeEvents(courses) {
 }
 
 function updateSchedule(courses) {
-    console.log("------------------------------------------")
-    console.log(JSON.stringify(courses));
-    console.log("------------------------------------------")
     myApp.courses = courses;
     clearOneTimeEvents();
     updateOneTimeEvents(courses);
@@ -998,14 +969,8 @@ function generateSchedule() {
 
     // Grab courses to schedule
     var courseIds = myApp.coursesToSchedule;
-    if (courseIds.length < 1) {
-        return;
-        // TODO error message
-    }
 
     var request = { "optimization": selectedValue, "courseIds": courseIds };
-
-    console.log("Payload: " + JSON.stringify(request));
 
     xhr.send(JSON.stringify(request));
 
@@ -1042,13 +1007,11 @@ function customEventCallout(name, days, startTime, endTime) {
 
     //TODO check valid input
 
-    console.log("Payload: " + JSON.stringify(send));
     xhr.send(JSON.stringify(send));
 
     // Generate schedule with new event
     xhr.onreadystatechange = function () {
         if (xhr.readyState == 4 && xhr.status == 200) {
-            console.log("Custom Event: " + JSON.stringify(xhr.responseText));
             var text = JSON.parse(xhr.responseText);
 
             //get course id to add to scheduleing
